+++ package.json
@@ -14,6 +14,7 @@
     "@babel/plugin-proposal-class-properties": "^7.0.0-beta.35",
     "@babel/preset-env": "^7.0.0-beta.35",
     "@babel/preset-stage-0": "^7.0.0-beta.35",
+    "ava": "^0.24.0",
     "babel-eslint": "^8.0.3",
     "babel-preset-flow": "^7.0.0-beta.3",
     "eslint": "^4.13.1",
