+++ src/index.js
@@ -1,0 +1,4 @@
+import Component from './Component'
+import createServer from './createServer'
+
+export default { Component, createServer }
